Dinesh Hackathon v1/ai-debate-partner/backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic

aiofiles
